backend/pyproject.toml
[project]
name = "qr-lifecycle-backend"
version = "0.1.0"
description = "QR lifecycle management backend"
readme = "../README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{name = "AutoGenerated", email = "dev@example.com"}]
dependencies = [
    "fastapi==0.110.0",
    "uvicorn[standard]==0.27.1",
    "SQLAlchemy==2.0.25",
    "alembic==1.13.1",
    "psycopg2-binary==2.9.9",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "python-multipart==0.0.6",
    "pydantic[email]==1.10.14",
    "redis==5.0.1",
    "Pillow==10.2.0",
    "qrcode==7.4.2",
    "numpy==1.26.4",
    "opencv-python-headless==4.9.0.80",
    "python-dotenv==1.0.1",
    "httpx==0.26.0",
]

[project.optional-dependencies]
dev = [
    "ruff==0.2.2",
    "black==24.2.0"
]

tests = [
    "pytest==7.4.4",
    "pytest-asyncio==0.23.5",
    "pytest-cov==4.1.0",
    "pytest-mock==3.12.0"
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["app"]

[tool.pytest.ini_options]
pythonpath = ["src"]
asyncio_mode = "auto"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "UP", "B"]
ignore = ["B008"]

